third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/injected-inline-script-blocked-expected.txt

CONSOLE ERROR: line 1: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 5: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:*". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.
